+++ data.php
@@ -1,0 +1,50 @@
+<?php
+$projects = [
+	'Все', 
+	'Входящие', 
+	'Учеба', 
+	'Работа', 
+	'Домашние дела', 
+	'Авто'
+];
+
+$tasks = [
+	[
+		'task' => 'Собеседование в IT компании',
+		'date' => '01.06.2018',
+		'category' => 'Работа',
+		'done' => false
+	],
+	[
+		'task' => 'Выполнить тестовое задание',
+		'date' => '25.05.2018',
+		'category' => 'Работа',
+		'done' => false
+	],
+		[
+		'task' => 'Сделать задание первого раздела',
+		'date' => '21.04.2018',
+		'category' => 'Учеба',
+		'done' => true
+	],
+		[
+		'task' => 'Встреча с другом',
+		'date' => '22.04.2018',
+		'category' => 'Входящие',
+		'done' => false
+	],
+		[
+		'task' => 'Купить корм для кота',
+		'date' => '',
+		'category' => 'Домашние дела',
+		'done' => false
+	],
+		[
+		'task' => 'Заказать пиццу',
+		'date' => '',
+		'category' => 'Домашние дела',
+		'done' => false
+	],
+];
+
+?>
