+++ templates/index.php
@@ -1,4 +1,21 @@
- <h2 class="content__main-heading">Список задач</h2>
+ <?php
+// показывать или нет выполненные задачи
+$show_complete_tasks = rand(0, 1);
+// устанавливаем часовой пояс в Московское время
+date_default_timezone_set('Europe/Moscow');
+
+$days = rand(-3, 3);
+$task_deadline_ts = strtotime("+" . $days . " day midnight"); // метка времени даты выполнения задачи
+$current_ts = strtotime('now midnight'); // текущая метка времени
+
+// запишите сюда дату выполнения задачи в формате дд.мм.гггг
+$date_deadline = date("d.m.Y", $task_deadline_ts);
+
+// в эту переменную запишите кол-во дней до даты задачи
+$days_until_deadline = floor(($task_deadline_ts - $current_ts) / 86400);
+?>
+
+<h2 class="content__main-heading">Список задач</h2>
 
 <form class="search-form" action="index.html" method="post">
     <input class="search-form__input" type="text" name="" value="" placeholder="Поиск по задачам">
